Add a test that always publishes in Debug
Fixes #99

Also move some tests over to reduce unnecessary deployment
