+++ test/FunctionalTests/PublishWithDebugTest_CoreCLR.cs
@@ -1,23 +1,18 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Collections.Generic;
-using System.IO;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Server.IntegrationTesting;
-using Microsoft.AspNetCore.Testing.xunit;
 using Microsoft.Extensions.Logging.Testing;
 using Xunit;
 using Xunit.Abstractions;
 
 namespace FunctionalTests
 {
-    [OSSkipCondition(OperatingSystems.Linux)]
-    [OSSkipCondition(OperatingSystems.MacOSX)]
-    public class ViewCompilationOptions_Desktop_ScenarioRefAssembliesDoNotGetPublished :
-        LoggedTest, IClassFixture<ViewCompilationOptions_Desktop_ScenarioRefAssembliesDoNotGetPublished.TestFixture>
+    public class PublishWithDebugTest_CoreCLR :
+        LoggedTest, IClassFixture<PublishWithDebugTest_CoreCLR.TestFixture>
     {
-        public ViewCompilationOptions_Desktop_ScenarioRefAssembliesDoNotGetPublished(
+        public PublishWithDebugTest_CoreCLR(
             TestFixture fixture,
             ITestOutputHelper output)
             : base(output)
@@ -27,26 +22,30 @@ namespace FunctionalTests
 
         public ApplicationTestFixture Fixture { get; }
 
-        [ConditionalFact]
-        public async Task PublishingWithOption_AllowsPublishingRefAssemblies()
+        [Fact]
+        public async Task PublishingInDebugWorks()
         {
             using (StartLog(out var loggerFactory))
             {
                 // Arrange
                 var deployment = await Fixture.CreateDeploymentAsync(loggerFactory);
 
-                // Act & Assert
-                Assert.True(Directory.Exists(Path.Combine(deployment.ContentRoot, "refs")));
+                // Act
+                var response = await deployment.HttpClient.GetStringWithRetryAsync(
+                    deployment.ApplicationBaseUri,
+                    loggerFactory.CreateLogger(Fixture.ApplicationName));
+
+                // Assert
+                TestEmbeddedResource.AssertContent("SimpleAppTest.Home.Index.txt", response);
             }
         }
 
-        public class TestFixture : DesktopApplicationTestFixture<SimpleApp.Startup>
+        public class TestFixture : CoreCLRApplicationTestFixture<SimpleApp.Startup>
         {
             protected override DeploymentParameters GetDeploymentParameters()
             {
                 var deploymentParameters = base.GetDeploymentParameters();
-                deploymentParameters.PublishEnvironmentVariables.Add(
-                    new KeyValuePair<string, string>("MvcRazorExcludeRefAssembliesFromPublish", "false"));
+                deploymentParameters.Configuration = "Debug";
 
                 return deploymentParameters;
             }
