+++ test/FunctionalTests/ViewCompilationOptionsTest_CoreCLR.cs
@@ -11,47 +11,6 @@ using Xunit.Abstractions;
 
 namespace FunctionalTests
 {
-    public class ViewCompilationOptions_CoreCLR :
-        LoggedTest, IClassFixture<CoreCLRApplicationTestFixture<SimpleApp.Startup>>
-    {
-        public ViewCompilationOptions_CoreCLR(
-            CoreCLRApplicationTestFixture<SimpleApp.Startup> fixture,
-            ITestOutputHelper output)
-            : base(output)
-        {
-            Fixture = fixture;
-        }
-
-        public ApplicationTestFixture Fixture { get; }
-
-        [Fact]
-        public async Task Precompilation_PreventsRefAssembliesFromBeingPublished()
-        {
-            using (StartLog(out var loggerFactory))
-            {
-                // Arrange
-                var deployment = await Fixture.CreateDeploymentAsync(loggerFactory);
-
-                // Act & Assert
-                Assert.False(Directory.Exists(Path.Combine(deployment.ContentRoot, "refs")));
-            }
-        }
-
-        [Fact]
-        public async Task Precompilation_PublishesPdbsToOutputDirectory()
-        {
-            using (StartLog(out var loggerFactory))
-            {
-                // Arrange
-                var deployment = await Fixture.CreateDeploymentAsync(loggerFactory);
-                var pdbPath = Path.Combine(deployment.ContentRoot, Fixture.ApplicationName + ".PrecompiledViews.pdb");
-
-                // Act & Assert
-                Assert.True(File.Exists(pdbPath), $"PDB at {pdbPath} was not found.");
-            }
-        }
-    }
-
     public class ViewCompilationOptions_CoreCLR_ScenarioRefAssembliesDoNotGetPublished :
         LoggedTest, IClassFixture<ViewCompilationOptions_CoreCLR_ScenarioRefAssembliesDoNotGetPublished.TestFixture>
     {
